src/djangoapidemo/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
click==8.0.4
cmake==3.22.4
debugpy==1.5.1
decorator==5.1.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
dlib==19.23.1
entrypoints==0.4
executing==0.8.2
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.9.1
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
Markdown==3.3.7
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
rel==0.4.7
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
wcwidth==0.2.5
zipp==3.8.0
